Invoke providers only if needed
This commit ensures only providers that provide dependencies used by injected components are invoked. If a provider isn't used then it won't be invoked.

In order to make this work, the injected component is invoked as a function so React.creatElement isn't called under the hood. This reason for this change is that this allows us to wrap props with a Proxy.