Invoke providers only if needed (#22)
* Dependencies from subgraphs were not resolved when injecting components

* Invoke providers only if needed

This commit ensures only providers that provide dependencies used by injected components are invoked. If a provider isn't used then it won't be invoked.

In order to make this work, the injected component is invoked as a function so React.creatElement isn't called under the hood. This reason for this change is that this allows us to wrap props with a Proxy.

* Test react lifecycle + update eslint

* Get rid of react-testing-library. Long live @testing-library/react

* Some work towards getting the example to work

* Don't lint dist folder

* Don't need this anymore :)

* lint

* remove unused code

Co-authored-by: Artal Druk <artald@wix.com>