Fix injection of memoized components
Disclaimer: I'm not 100% certain about this commit.

The issue happens because Obsidian invokes components as functions which is frown upon. We should be using React.createElement, but we can't because it seems like it spreads props which breaks our proxied props useless.
This PR attempts to check if a component is memoized, and renders the concrete component.

Fixes #60