Execute mediatorObservable.addSource callbacks immediately to create an initial value (#115)
MediatorObservable relies on other observables to produce a data stream. It's used to either merge two observables into a new observable or to "map" a single observable into a new kind. For these reasons, MediatorObservables should always have an initial value that's based on the mediated observables.

The current method of providing an initial value causes code duplication since it requires the developer to execute the logic contained in the `onChange` callbacks.

This PR change MediatorObservable so it executes the `onNext` callback immediately when `addSource` is invoked. This will effectively create the initial value, without needing to pass it explicitly in the constructor.

Co-authored-by: Guy Carmeli <>