[BREAKING] Use @LifecycleGraphs as service locators only if they are created (#121)
This PR addresses a case where `@LifecycleBound` graphs were leaking if they were used as service locators.

From now on, `@LifecycleBound` graphs can still be used as service locators, but only after they were created by a "lifecycle owner" - a React component or hook. This change ensures that the graph is properly released when the React construct is unmounted.