Add ability to map multiple observables in a single callback (#127)
A common case with MediatorObservables is the need to create a side effect based on multiple Observables. The mapSource function maps a single source. When it's used to map multiple sources, in order to adhere to the DRY principle we won't be able to inline to onNext callback as it needs to be shared in multiple calls to mapSource.

This PR adds the mapSources function which receives multiple sources and consequently makes it easier to share the same callback.