Fix file path resolving in ESLint rule
The path of nested subgraphs wasn't resolved properly as it relayed on the path of the file being linted. This PR changes the logic slightly so the path is calculated based on the path of the initial graph being linted.

This PR also fixes a few NPE's found while linting an actual project. Apparently the types exported by TSESTree aren't 100% accurate.