Introduce lifecycle scopes (#167)
### Background
Complex screens that display many components generally tend to have big graphs. In such cases, splitting the graph into multiple smaller subgraphs where each subgraph accompanies a single component can be beneficial. Graphs that are bound to a component should behave slightly differently than graphs that are bound to the lifecycle of a feature. For example, when a component is displayed in multiple screens, a new graph should be created for each component instance.

This PR adds the ability to change the scope of lifecycle-bound graphs to a single component and its direct children.

### API
The `@LifecycleBound` decorator now accepts a new `scope` option. The `scope` option can be set to either `feature` or `component`. The default value is `feature`. When the `scope` option is set to `component`, a new graph is created for each component instance. The same graph is shared between the component and its children.

```ts
import {LifecycleBound, Graph, ObjectGraph} from 'react-obsidian';

@LifecycleBound({ scope: 'component' }) @Graph()
class ScopedLifecycleBoundGraph extends ObjectGraph {
}
```
### Example
First, let's create a graph that is bound to the lifecycle of a component. The graph provides a single dependency `text`.

```tsx
import {LifecycleBound, Graph, ObjectGraph, injectComponent, Provides} from 'react-obsidian';

@LifecycleBound({ scope: 'component' }) @Graph()
class ScopedGraph extends ObjectGraph<ParentProps> {
  constructor(private props: ParentProps) {
    super(props);
  }

  @Provides()
  text() {
    return this.props.text;
  }
}
```

Next, let's create two components that use the graph. The Parent component renders a Child component that's also injected with the `ScopedGraph` we created above.

```tsx
type ParentProps = {
  text: string;
};

const Parent = injectComponent<ParentProps>(() => {
  return <Child />;
}, ScopedLifecycleBoundGraph);

type ChildProps = DependenciesOf<ScopedLifecycleBoundGraph, 'text'>

const Child = injectComponent(({text}: ChildProps) => {
  return <p>{text}</p>;
}, ScopedLifecycleBoundGraph);
```

Finally, let's render the Parent component with different texts.

```tsx
function App() {
  return (
    <div>
      <Parent text="Hello" />
      <Parent text="World" />
    </div>
  );
}
```

When the `scope` option is set to `component`, the `ScopedGraph` is created **twice**, once for each `Parent` component. The `Child` component is injected with the `ScopedGraph` instance that's bound to the `Parent` component.