Update to TypeScript 5.5 + ESLint 9.9 (#171)
This PR updates TypeScript to v5.5.x and ESLint to 9.9.x. As part of the ESLint upgrade, several ESLing packages were dropped since they don't support the new flat config.

This PR is a prerequisite to evaluating if we can drop the requirement of `@babel/plugin-proposal-decorators` since native support for [decorators](https://devblogs.microsoft.com/typescript/announcing-typescript-5-0/#decorators) in TS was advanced to the stage 3 implementation.